fork chainmanager
