+++ state/factory/factory.go
@@ -88,6 +88,7 @@ type (
 		PutBlock(context.Context, *block.Block) error
 		WorkingSet(context.Context) (protocol.StateManager, error)
 		WorkingSetAtHeight(context.Context, uint64, ...*action.SealedEnvelope) (protocol.StateManager, error)
+		StateReaderAt(header *block.Header) (protocol.StateReader, error)
 	}
 
 	// factory implements StateFactory interface, tracks changes to account/contract and batch-commits to DB
@@ -537,6 +538,10 @@ func (sf *factory) States(opts ...protocol.StateOption) (uint64, state.Iterator,
 // ReadView reads the view
 func (sf *factory) ReadView(name string) (interface{}, error) {
 	return sf.protocolView.Read(name)
+}
+
+func (sf *factory) StateReaderAt(header *block.Header) (protocol.StateReader, error) {
+	panic("implement me")
 }
 
 //======================================
